+++ frameworks/core/pipeline/pipeline_context.cpp
@@ -2011,7 +2011,7 @@ void PipelineContext::OnIdle(int64_t deadline)
     FlushPageUpdateTasks();
 }
 
-void PipelineContext::OnVirtualKeyboardHeightChange(double keyboardHeight)
+void PipelineContext::OnVirtualKeyboardHeightChange(float keyboardHeight)
 {
     CHECK_RUN_ON(UI);
     double positionY = 0;
